refactor: move notionImageService into notro-loader/image-service
- Add packages/notro-loader/image-service.ts as a new export entry point
- Export `notionImageService` config object (replaces notionImageServiceConfig() fn)
- entrypoint is now "notro-loader/image-service" — self-documenting in astro.config.mjs
- Remove duplicate notionImageService.ts from both templates (blank and blog)
- Remove now-empty templates/blank/src/lib/ directory
